[Pods] Remove orphaned references and let CocoaPods reconfigure.
It appears that, for whatever reason, there were some Xcode file
references that were orphaned and those were causing CocoaPods to error
out. Opening and building with Xcode.app still worked, because Xcode.app
tries to fix any issues it finds.

It also appears that the project was manually configured to also link
the pods into the GBA4iOS-Simulator target, which I undid and let
CocoaPods deal with that instead. The only real difference I found
between the two targets was that the sim target set a CPP definition of
`SIMULATOR=1`, which I've made sure is maintained.