Convert to Dropbox API V2
Mostly a direct translation from v1 to v2 function calls. API functionality was also switched from a delegate system to a callback system. On the first app launch with this version, files containing serialized v1 objects are deleted and the Dropbox user is logged out. The user should manually reauthenticate.